@@ -18,8 +18,8 @@
     "@styled-icons/material-outlined": "^10.34.0",
     "dotenv-webpack": "^5.1.0",
     "lodash": "^4.17.21",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "19",
+    "react-dom": "19",
     "react-is": "16.11",
     "react-router-dom": "6.2.1",
     "react-syntax-highlighter": "^15.5.0",
@@ -30,6 +30,8 @@
     "@babel/plugin-transform-runtime": "^7.16.5",
     "@babel/preset-env": "^7.16.5",
     "@babel/preset-react": "^7.16.5",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
     "babel-loader": "^8.2.3",
     "css-loader": "6.5.1",
     "html-webpack-plugin": "^5.5.0",
