@@ -15,12 +15,10 @@
 // This file is a top-level boilerplate. It ties the JS in to the .html document.
 
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import App from './App';
 
-
-ReactDOM.render(
-  <App />
-  , document.getElementById('app')
-);
+const container = document.getElementById('app');
+const root = createRoot(container);
+root.render(<App />);
