@@ -54,12 +54,10 @@ module.exports = {
           loader: "babel-loader",
           options: {
             "presets": [
-              "@babel/preset-env", 
-              "@babel/preset-react",
-              {
-                'plugins': ["@babel/plugin-transform-runtime", '@babel/plugin-proposal-class-properties']
-              }
-            ]
+              "@babel/preset-env",
+              ["@babel/preset-react", { "runtime": "automatic" }]
+            ],
+            "plugins": ["@babel/plugin-transform-runtime", "@babel/plugin-proposal-class-properties"]
           }
         }
       },
